+++ tests/test_utils.py
@@ -1,6 +1,8 @@
 import unittest
+
 from puzzle.board import Board
 from puzzle.utils import generate_multiple_boards
+
 
 class TestGenerateMultipleBoards(unittest.TestCase):
 
@@ -27,6 +29,7 @@ class TestGenerateMultipleBoards(unittest.TestCase):
         self.assertEqual(len(boards), count)
         for board in boards:
             self.assertIsInstance(board, Board)
+
 
 if __name__ == '__main__':
     unittest.main()
